+++ src/features/members/components/NewMemberForm/index.tsx
@@ -1,5 +1,5 @@
 import LoadingModal from '@/components/LoadingModal';
-import MemberForm from '@/features/members/components/MemberForm';
+import MemberForm from '@/features/members/components/BaseMemberForm';
 import axios from 'axios';
 import { useRouter } from 'next/router';
 import toast from 'react-hot-toast';
